+++ src/components/shared/ComponentDecorators.js
@@ -3,15 +3,15 @@ import React from 'react'
 
 function getWrappedComponetDisplayName(WrappedComponent) {
   return WrappedComponent.displayName || WrappedComponent.name || 'Component'
-} 
+}
 
 /**
  * Add classNames object to props
- * 
+ *
  * @export
  * @param {object} WrappedComponent - The React component to be wrapped
  * @param {object} styles - The classNames object translated from imported scss/css file
- * @returns 
+ * @returns
  */
 export function addStylesToPropsDecorator(WrappedComponent, styles) {
   return class WithStyles extends React.Component {
@@ -25,10 +25,10 @@ export function addStylesToPropsDecorator(WrappedComponent, styles) {
 
 /**
  * Wrap a componet with a div with specific styles
- * 
+ *
  * @param {object} WrappedComponent - The React component to be wrapped
  * @param {object} wrappertStyles  - The styles assined to wrapper div
- * @returns 
+ * @returns
  */
 export function addStyledWrapperDecorator(WrappedComponent, wrappertStyles) {
   return class WrapWithStyledDiv extends React.Component {