+++ webpack.config.js
@@ -71,10 +71,6 @@ const webpackConfig = {
         use: 'babel-loader'
       },
       {
-        test: /node_modules\/react-flex-carousel\/index.js/,
-        use: 'babel-loader'
-      },
-      {
         test: /\.(scss|sass)$/,
         use: isProduction ? ExtractTextPlugin.extract({
           fallback: 'style-loader',
@@ -126,6 +122,8 @@ const webpackConfig = {
           use: [{
             loader: 'css-loader',
             options: {
+              minimize: isProduction,
+              sourceMap: !isProduction,
               importLoaders: 1
             }
           }, {